refactor: rename review.types to queue.types
